feat

+++ app/Http/Controllers/Varieties/DeleteVarietyController.php
@@ -10,6 +10,12 @@ class DeleteVarietyController extends Controller
 {
     public function __invoke(Variety $variety)
     {
+        if ($variety->hasAssociatedRecords()) {
+            return back()->withErrors([
+                'variety' => 'No se puede eliminar la variedad porque tiene registros asociados (centros de costo, evaluaciones de proyecto u otros). Elimine primero los registros relacionados.'
+            ]);
+        }
+
         $variety->delete();
     }
 }
